Refactored Navigation code into a separate folder/file that will be used to handle all of the navigation stuff. Going to work on getting a bottom tab navigator working soon soon soon.
